Crate unify_attribute_poll

This library features an engine to poll reported values of attributes with a given time interval. The Poll Engine determines if an attribute shall be polled based on a timeout for the given attribute, in which case it clears the reported value of the attribute. The Attribute Resolver then takes over and tries to resolve the attribute.

Configuration

The Poll Engine is configurable using the PollEngineConfig and supports:

Backoff: The minimum time between 2 polls.

Max Polls Per Hour: Maximum number of polls allowed during 1 hour.

Default Poll Interval: In case an attribute is registered for polling with no interval, this interval is used.

AttributePoll

To control the PollEngine the AttributePoll object is used. More detailed information can be found in the trait documentation of the AttributePollTrait

C interface

this unify_attribute_poll module exposes an C interface. see attribute_poll.h for more information.

Example

let attribute_poll = AttributePoll::default();
// Create "dummy" attribute for sake of the example.
let my_attribute = Attribute::new_from_handle(42).unwrap();
// Register my_attribute to be polled every 30 seconds
attribute_poll.register(my_attribute.clone(), 30);
// Restart the poll
attribute_poll.restart(my_attribute.clone());
// Stop polling the attribute
attribute_poll.deregister(my_attribute.clone());

Modules

__mock_MockAttributePollTrait

__mock_MockAttributePollTrait_AttributePollTrait

Structs

AttributePoll

Its purpose is to interface with the single Attribute PollEngine instance. And allows for easy sharing of the PollEngine,

MockAttributePollTrait

Interface for the Attribute Poll engine.

PollEngineConfig

Configuration for the PollEngine

Traits

AttributePollTrait

Interface for the Attribute Poll engine.