Trait unify_attribute_poll::attribute_poll_trait::AttributePollTrait

pub trait AttributePollTrait {
    fn initialize(&self, config: PollEngineConfig);

    fn register(&self, attribute: Attribute, interval: IntervalType);

    fn deregister(&self, attribute: Attribute);

    fn schedule_now(&self, attribute: Attribute);

    fn restart(&self, attribute: Attribute);

    fn pause(&self);

    fn resume(&self);
}

This traits acts as a facade for the [PollEngine] which runs the actual attribute polling logic. There can be multiple instances of the objects implementing this trait but they will always control only one unique [PollEngine]. Objects implementing this trait are therefore cheap to construct.

Required methods

fn initialize(&self, config: PollEngineConfig)

starts the Poll Engine and puts its state to running. e.g. resume is not required to be explicitly called. Initialize will return immediately when its called more as once. With the configuration file the poll behavior can be influenced; The rules for timer scheduling are:

If deadline of next entry < last_poll + backoff, queue timeout for in backoff seconds,

else queue timeout for deadline of next entry.

Arguments

config persists as long as the application is alive. Config cannot be re-submitted by calling the .initialize() twice

fn register(&self, attribute: Attribute, interval: IntervalType)

Adds a [Attribute] to the Poll Engine with a given interval. Internally it will recalculate the next deadline and reschedule poll timeouts accordingly.

Arguments

attribute a valid attribute corresponding a zwave cluster property

interval interval is in seconds and should be relative from time now

fn deregister(&self, attribute: Attribute)

removes an attribute from the Poll Engine. Does nothing if the entry does not exist.

Arguments

attribute a valid attribute corresponding a zwave cluster property. Attribute must exists in the poll engine

fn schedule_now(&self, attribute: Attribute)

registers the attribute at the front of the queue. The attribute will be polled on a first available moment.

Arguments

attribute a valid attribute corresponding a zwave cluster property

fn restart(&self, attribute: Attribute)

re-inserts the attribute to the poll queue with the interval it was given at time of registering. time of poll = current time + interval.

Arguments

attribute a valid attribute corresponding a zwave cluster property

fn pause(&self)

no attributes will be polled. When the attribute poller is paused it can still receive other commands.

Note that, depending on the configured back-off time, it can take a longer time for the attribute poller to catch up polling attributes if the attribute poller is paused for a longer period of time. This means that attributes will be polled later as expected.

fn resume(&self)

Resume polling attribute