Trait unify_attribute_poll::AttributePollTrait

pub trait AttributePollTrait {
    fn initialize(&self, config: PollEngineConfig);

    fn register(&self, attribute: Attribute, interval: u32);

    fn deregister(&self, attribute: Attribute);

    fn queue(&self, attribute: Attribute);

    fn restart(&self, attribute: Attribute);

    fn pause(&self);

    fn resume(&self);
}

Interface for the Attribute Poll engine.

Required methods

fn initialize(&self, config: PollEngineConfig)

initializes the Poll Engine. Secondarily it will set the state of to running. e.g. resume is not required to be explicitly called. With the configuration file the poll behavior can be influenced; The rules for timer scheduling are:

If deadline of next entry < last_poll + backoff, queue timeout for in backoff seconds,

else queue timeout for deadline of next entry.

fn register(&self, attribute: Attribute, interval: u32)

Adds a [Attribute] to the Poll Engine with a given interval. Internally it will recalculate the next deadline and reschedule poll timeouts accordingly

fn deregister(&self, attribute: Attribute)

removes an attribute from the Poll Engine. Does nothing if the entry does not exist.

fn queue(&self, attribute: Attribute)

similar to register, only that no interval is given, meaning the attribute will be polled on the first available moment

fn restart(&self, attribute: Attribute)

re-inserts the attribute at the current time to the poll queue with the interval it was given at time of registering.

fn pause(&self)

no attributes will be polled. When the Poll Engine is paused it can still receive other commands

fn resume(&self)

attributes will be polled again.